solved

+++ Math/Problem1/main.cpp
@@ -3,26 +3,26 @@
 
 using namespace std;
 
-long long int sum1toNum( long long int input )
+unsigned long long sum1toNum( unsigned long long input )
 {
     return (1 + input) * (input/2) + ((input%2)*(input/2+1));
 }
 
-long long int sumUnderMultiples( long long int max , long long int by)
+unsigned long long sumUnderMultiples( unsigned long long max , unsigned long long by)
 {
     return sum1toNum( max / by ) * by ;
 }
 
 int main(int argc, char* argv[])
 {
-    long long int input,sum;
+    unsigned long long input,sum;
 
     if ( argc != 2 )
     {
         return -1;
     }
 
-    input = atoi(argv[1]);
+    input = atol(argv[1]);
 
     sum = sumUnderMultiples( input, 5 ) + sumUnderMultiples( input, 3 ) - sumUnderMultiples( input, 15 );
     cout << "Result : " << sum << endl;
